+++ ChangeLog.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Improve Deserialize usages by adding global `Deserialize`
 
 ### Changed
 - Replace `curl` and `xmlrpc-c` with `cpp-httplib`, `tinyxml2` and a custom `xmlrpc` client